move browser modes so they can be accessed in electron.js
